charges exist chec on delete

+++ app/Http/Controllers/Api/CompanyController.php
@@ -64,6 +64,7 @@ class CompanyController extends Controller
 
     public function destroy(CompanyRequest $request, $id)
     {
+        if (!$this->companyRepository->exist($id)) throw InvalidException::forInvalid("No existe");
 
         try {
             $this->companyRepository->delete($id);
